inject: split massive file (#28792)
Currently we have a big inject.go file serving many tasks. This splits
it up into validation, templating, and the core injection.

This will be useful to make a future PR I have not so big. This PR is
strictly moving code around, there should be no functional changes
